Refactored GraphQL API and broke down into use-case-specific files. This closes issue #76.
