+++ api/schema/entity.graphql
@@ -1,0 +1,28 @@
+type Entity {
+  id: String!
+  name: String!
+  imageUrl: String!
+  entityType: String!
+  mediaItems(page: Int, limit: Int): MediaItemConnection!
+}
+
+type EntityItemConnection {
+  nodes: [Entity!]
+  totalCount: Int!
+}
+
+type ExploreResponse {
+  people: [Entity!]
+  places: [Entity!]
+  things: [Entity!]
+}
+
+extend type Query {
+  explore: ExploreResponse!
+  entities(entityType: String!, page: Int, limit: Int): EntityItemConnection!
+  entity(id: String!): Entity!
+}
+
+extend type Mutation {
+  updateEntity(id: String!, name: String!): Boolean!
+}
