+++ api/schema/schema.graphql
@@ -1,0 +1,17 @@
+scalar Time
+
+type AutocompleteResponse {
+  id: String!
+  name: String!
+}
+
+extend type Query {
+  search(q: String, id: String, page: Int, limit: Int): MediaItemConnection!
+  autocomplete(q: String!): [AutocompleteResponse]
+  favourites(page: Int, limit: Int): MediaItemConnection!
+}
+
+extend type Mutation {
+  upload(file: Upload!): Boolean!
+  updateFavourite(id: String!, type: String!): Boolean!
+}
